Add touch event to send Comparison choice to a server
This adds very basic handling of tapping an image, posting a new Comparison record to the RaresServer.

Next step should be to get the Redux data flow and action pattern in place, so that image urls are stored in Redux state and a reload action can been applied as the callback after a Comparison is successfully submitted.